Refactor of model selection feature on form and addition of automated model list search.
The method of selecting the GPD theoretical parametrization and GPD selection has been changed to allow the user to select from a list of theoretical models. All models are stored in the data/models/ directory. Currently the models are hardcoded but a function to find and return them in an automated manner has been added and now needs to be implemented. Users are now able to select which GPD they want to plot: GPD_H or GPD_E. This is done via radio buttons that have been added to the front page.

Summary of changes:

- Refactored location of models to data/models/model_<user-model-name>/
- Added radio button to select GPD of interest.
- Implemented new buttons into html and model_plot( ... ) so that the csv file that is opened points to the users selection.
- Added model_search() function that returns a list of all model folders found in model folder. This needs to be implemented into form.py so that the model selection menu is automatically filled.