Major update:
Summary of changes:
-- Added dynamic form filling
-- Modified the column divisions a bit.

The values for dropdown boxes associated with xbj and t are now filled dynamically according to the selected model and gpd values. This is done by grabbing the unique data frame values from the model on the local drive. In the log run this should be done by pulling limits from the redis database. Changes included a new params function in routes.py and more javascript than I ever want to write again in the main html.

Please hire a frontend dev Simonetta : )

Bugs:

-- For some reason I get a error "invalid parameter" if I hit enter after filling out the Q2 box. Everything works fine I instead hit plot. I think this is a remnant of switching from a DecimalField to SelectField in Flask. Hopefully Ruoshi can point out the origin.

Next step is to clean up the HTML and add our logo along with the CNF and UVA.